filter

+++ VolunteerAdmin/Controllers/VolunteersController.cs
@@ -20,9 +20,35 @@ namespace VolunteerAdmin.Controllers
         }
 
         // GET: Volunteers
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string filter)
         {
-            return View(await _context.Volunteers.ToListAsync());
+            var volunteers = from s in _context.Volunteers select s;
+            //a.	Approved/Pending Approval (volunteers in both approval statuses are included)
+            //            b.Approved
+            //c.Pending Approval
+            //d.Disapproved
+            //e.Inactive
+            //f.All
+            //var search =>
+
+            //switch (filter)
+            //{
+            //    case "Approved":
+            //     {
+
+            //            break;
+            //     }
+            //}
+            //if (filter != null)
+            //    return View("Delete");
+            bool filterBool = filter == "Approved";
+            if (!String.IsNullOrEmpty(filter))
+            {
+                volunteers = volunteers.Where(v => v.Approved==filterBool);
+            }
+
+
+            return View(await volunteers.ToListAsync());
         }
 
         // GET: Volunteers/Details/5
@@ -56,13 +82,22 @@ namespace VolunteerAdmin.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,FirstName,LastName,Username,Password,HomePhone,CellPhone,WorkPhone,Address,Email,Education,EmergencyName,EmergencyHomePhone,EmergencyWorkPhone,EmergencyEmail,EmergencyAddress,DLCopyOnFile,SSCopyOnFile,Approved")] Volunteer volunteer)
+        public async Task<IActionResult> Create([Bind("FirstName,LastName,Username,Password,HomePhone,CellPhone,WorkPhone,Address,Email,Education,EmergencyName,EmergencyHomePhone,EmergencyWorkPhone,EmergencyEmail,EmergencyAddress,DLCopyOnFile,SSCopyOnFile,Approved")] Volunteer volunteer)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(volunteer);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(volunteer);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException)
+            {
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the program persists " +
+                    "see your system administrator.");
             }
             return View(volunteer);
         }
@@ -88,38 +123,76 @@ namespace VolunteerAdmin.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,FirstName,LastName,Username,Password,HomePhone,CellPhone,WorkPhone,Address,Email,Education,EmergencyName,EmergencyHomePhone,EmergencyWorkPhone,EmergencyEmail,EmergencyAddress,DLCopyOnFile,SSCopyOnFile,Approved")] Volunteer volunteer)
+        public async Task<IActionResult> Edit(int? id, [Bind("ID,FirstName,LastName,Username,Password,HomePhone,CellPhone,WorkPhone,Address,Email,Education,EmergencyName,EmergencyHomePhone,EmergencyWorkPhone,EmergencyEmail,EmergencyAddress,DLCopyOnFile,SSCopyOnFile,Approved")] Volunteer volunteer)
         {
-            if (id != volunteer.ID)
+            //if (id != volunteer.ID)
+            //{
+            //    return NotFound();
+            //}
+
+            //if (ModelState.IsValid)
+            //{
+            //    try
+            //    {
+            //        _context.Update(volunteer);
+            //        await _context.SaveChangesAsync();
+            //    }
+            //    catch (DbUpdateConcurrencyException)
+            //    {
+            //        if (!VolunteerExists(volunteer.ID))
+            //        {
+            //            return NotFound();
+            //        }
+            //        else
+            //        {
+            //            throw;
+            //        }
+            //    }
+            //    return RedirectToAction(nameof(Index));
+            //}
+            //return View(volunteer);
+            if (id == null)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            var volunteerToUpdate = await _context.Volunteers.FirstOrDefaultAsync(v => v.ID == id);
+            if (await TryUpdateModelAsync<Volunteer>(volunteerToUpdate,
+                "",
+                v => v.FirstName,
+                v => v.LastName,
+                v => v.Username,
+                v => v.Password,
+                v => v.HomePhone,
+                v => v.CellPhone,
+                v => v.WorkPhone,
+                v => v.Address,
+                v => v.Email,
+                v => v.Education,
+                v => v.EmergencyName,
+                v => v.EmergencyHomePhone,
+                v => v.EmergencyWorkPhone,
+                v => v.EmergencyEmail,
+                v => v.EmergencyAddress,
+                v => v.DLCopyOnFile,
+                v => v.SSCopyOnFile,
+                v => v.Approved))
             {
                 try
                 {
-                    _context.Update(volunteer);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException)
                 {
-                    if (!VolunteerExists(volunteer.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Unable to save changes." +
+                        "Try again, and if the problem persists, see your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(volunteer);
+            return View(volunteerToUpdate);
         }
 
         // GET: Volunteers/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -127,12 +200,19 @@ namespace VolunteerAdmin.Controllers
             }
 
             var volunteer = await _context.Volunteers
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (volunteer == null)
             {
                 return NotFound();
             }
 
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
+            }
             return View(volunteer);
         }
 
@@ -142,9 +222,25 @@ namespace VolunteerAdmin.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var volunteer = await _context.Volunteers.FindAsync(id);
-            _context.Volunteers.Remove(volunteer);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            //_context.Volunteers.Remove(volunteer);
+            //await _context.SaveChangesAsync();
+            //return RedirectToAction(nameof(Index));
+
+            if (volunteer == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Volunteers.Remove(volunteer);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException)
+            {
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool VolunteerExists(int id)
